Integrate from latest book source
Big changes:
Switch to PathIntegrator for default Integrator
Switch to HaltonSampler for default Sampler
Switch to 16spp for default number of pixel samples.
Fix uninitialized memory bug in InitProfiler()

Smaller changes:
Add profiler category for camera ray generation.
Many naming cleanups, as usual.
Return exit pupil bounds from RealisticCamera::SampleExitPupil() to make
  ray weight computation more straightforward.
Use unique_ptr for Mapping[23]D in Texture implementations

+++ src/lights/distant.cpp
@@ -40,14 +40,14 @@
 // DistantLight Method Definitions
 DistantLight::DistantLight(const Transform &LightToWorld, const Spectrum &L,
                            const Vector3f &wLight)
-    : Light(LightFlags::DeltaDirection, LightToWorld, MediumInterface()),
+    : Light((int)LightFlags::DeltaDirection, LightToWorld, MediumInterface()),
       L(L),
       wLight(Normalize(LightToWorld(wLight))) {}
 Spectrum DistantLight::Sample_Li(const Interaction &ref, const Point2f &u,
                                  Vector3f *wi, Float *pdf,
                                  VisibilityTester *vis) const {
     *wi = wLight;
-    *pdf = 1.f;
+    *pdf = 1;
     Point3f pOutside = ref.p + wLight * (2 * worldRadius);
     *vis =
         VisibilityTester(ref, Interaction(pOutside, ref.time, mediumInterface));
