+++ src/lights/infinite.cpp
@@ -42,7 +42,8 @@
 InfiniteAreaLight::InfiniteAreaLight(const Transform &LightToWorld,
                                      const Spectrum &L, int nSamples,
                                      const std::string &texmap)
-    : Light(LightFlags::Infinite, LightToWorld, MediumInterface(), nSamples) {
+    : Light((int)LightFlags::Infinite, LightToWorld, MediumInterface(),
+            nSamples) {
     // Read texel data from _texmap_ and initialize _Lmap_
     Point2i resolution;
     std::unique_ptr<RGBSpectrum[]> texels(nullptr);
